requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
playwright>=1.40
beautifulsoup4>=4.12
dnspython>=2.4
aiosmtplib>=3.0
aiosqlite>=0.19
pydantic>=2.5
pytz>=2024.1
